Refactor: 백엔드 코드 구조
