+++ BackEnd/src/routes/home.js
@@ -3,10 +3,9 @@
 const express = require('express');
 const router = express.Router();
 
-const ctrl = require('./service');
+const ctrl = require('../controllers/home');
 
 router.get("/", ctrl.homeView);
-
 router.get("/about", ctrl.aboutView);
 
 module.exports = router;
