+++ BackEnd/src/routes/index.js
@@ -3,14 +3,14 @@
 const express = require('express');
 const router = express.Router();
 
-const homeRoute = require('../controllers/home/route');
-const userRoute = require('../controllers/user/route');
-const boardRoute = require('../controllers/board/route');
+const homeRoute = require('./home');
+const userRoute = require('./user');
+const boardRoute = require('./board');
 
 const { getPostQueryString } = require('../functions/util');
 
 router.use("/", homeRoute);
 router.use("/user", userRoute);
-router.use("/board", getPostQueryString ,boardRoute);
+router.use("/board", getPostQueryString, boardRoute);
 
 module.exports = router;
